pyproject.toml
[project]
name = "loopfactory"
version = "0.1.0"
description = "LoopFactory - Multi-agent coordination and orchestration platform"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "apscheduler>=3.10.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"
